more type checking, remove STRING as a valid arg to CONTAINS?
considering how easy it is to use #I("1.2.3.4") or #I(ip-string), it's not too
onerous to require IP-LIKEs to CONTAINS?. all the methods that depend on
CONTAINS? only accept IP-LIKEs as well, so this makes things more consistent.
this diverges from python's netaddr library, but considering the IP syntax is so
light i don't think it's that big of a deal.